fix: more robust cookie management for plugins requiring launch
Instead of storing a cookie for every plugin that requires launch, we
create a single cookie with '+' separated value. We make use of the
cookieStore native API (available everywhere since June 2025) to access
cookie data. The variables are renamed to be more explicit. We now use
class-based SCSS for styling, instead of manually setting style.display
attribute.